Start using zipper buffer datastructure
The cursor’s x-position is broken (can’t move right), but scrolling up
and down when you reach the edges of the screen works well. At the moment
I’m getting an address boundary error if you scroll past the last line –
this is probably me not checking a condition correctly.