Allow two buffers in the editor
After thinking about it, multiple panes visiting the same file seems like a much
harder problem than multiple panes visiting different files. How do I
co-ordinate the file state between the two panes?

So, I’ve decided to instead go after editing multiple files (having multiple
buffers). First step is to change all of the buffer-touching functions to take
arguments instead of mutating the global editor state. Once I’ve done that, a
good structure for this buffer data will become clear (maybe to go in an
~EditorBuffer~ type?).

Note: halfway through I decided to just make that type, which I’ve called
~FileData~.

It is now possible (and currently required :sad:) to open two files at once in
the editor, and display them side by side. The scrolling is still a bit broken,
and though it’s possible to switch between the panes, the (visual) cusor always
stays in the left pane. (It is possible to scroll and edit the right-hand pane
once you have switched to it.) Interestingly, the scrolling is not broken for
the right-hand pane, which suggests I have done something silly. It looks like a
variable modification is being made twice. Though, it’s still not the behaviour
I want (and had), where the file only scrolls if the cursor is offscreen after
the move.

Next steps:
- fix the scrolling
- move the visual cursor to the correct pane
- fix drawing logic so if one pane is shorter than the screen, the other isn’t
  cut off at that point
- allow an arbitrary number of buffers
- allow an arbitrary number of panes
- allow arbitrary pane configurations
- make editing functions pure, returning a description of the edit instead of
  mutating
  - this one sounds fun, just might do this first
  - might replace the undo system with a tree of events instead of the current
    “keep a reference to an old version of a persistent datastructure” system