Fix: don’t clear rest of line on newline
It turns out this is why I was missing a space at the end of each line:
I think Terminal was including the current cursor position in what was
deleted, and so removing the final character of each line.

Deleting it fixes that problem, and doesn’t appear to introduce any
others: why was this necessary at all? Maybe in other terminals this is
required to overwrite the lines instead of adding new ones? I’ll leave
it out for now and see if it comes up again.