Add org-publish project for documentation
This commit moves the documentation to its own folder, and adds an
emacs-lisp file that publishes the documentation (to convert it to html
files using org-mode).

I’d really like to have the includes use search strings rather than line
numbers, but I can’t get the publishing to do it, either in links or in
includes. I’ll try again at some point, but this’ll have to do for now.